requirements.txt
pantilthat==0.0.5
paramiko==2.7.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
picraft==1.0
piglow==1.2.4
pigpio==1.38
Pillow==4.0.0
psutil==5.7.3
pyasn1==0.1.9
pycparser==2.20
pycrypto==2.6.1
PyJWT==1.4.2
pyOpenSSL==16.2.0
pyserial==3.2.1
pytz==2018.7
pyxdg==0.25
qrcode==6.0
rainbowhat==0.1.0
requests==2.12.4
requests-oauthlib==0.7.0
RPi.GPIO==0.7.0
sense-emu==1.0
sense-hat==2.2.0
simplejson==3.10.0
six==1.10.0
sn3218==1.2.7
spidev==3.4
sqlit==0.1.6
touchphat==0.0.1
twython==3.4.0
unicodecsv==0.14.1
unicornhathd==0.0.3
urllib3==1.19.1
websocket-client==0.57.0
websocket-server==0.4
Werkzeug==0.11.15
zope.event==4.5.0
zope.interface==5.1.2
chardet==4.0.0
paho-mqtt==1.6.1
baidu-aip==2.2.11.0
opencv-python==4.3.0.38
chardet==4.0.0